refactor(map-calibration): lift detection algorithms + introduce Detection DI extension (Phase C)
- Lift src/Mithril.MapCalibration/Detection/* into src/Mithril.MapCalibration.Detection/
- Lift FeatureMatchingRefiner + ORB descriptor cache (CachedOrbDescriptorProvider,
  OrbDescriptorWriter, OrbDescriptorManifest) from .Capture into .Detection
- Introduce AddMithrilMapCalibrationDetection in
  Mithril.MapCalibration.Detection.DependencyInjection, owning the OpenCv-tied
  registrations (IMapRegionRefiner / FeatureMatchingRefiner,
  MapCalibrationLocateOptions, CachedOrbDescriptorProvider, OrbDescriptorWriter)
  plus the ComputeOrbParamsHash helper
- Slim CaptureServiceCollectionExtensions: replace AddMithrilMapCalibrationEngine
  call with AddMithrilMapCalibrationDetection; delete moved registrations; keep
  the GameConfig-wired ICalibrationConfidenceGate override
- Delete AddMithrilMapCalibrationEngine from MapCalibrationServiceCollectionExtensions

Pragmatic architectural call (deviation from strict spec):
- Four boundary interfaces (IBaseTextureProvider, ICalibrationDetector,
  IIconTemplateProvider, IMapRegionRefiner) physically moved to the .Detection
  project but kept their Mithril.MapCalibration namespace. Their method
  signatures reference Detection-tier types (GrayImage, IconTemplate, etc.) that
  can't live in contracts without dragging algorithm machinery up. Namespace
  stability preserves the consumer API surface; project location follows the
  dependency direction.
- Same shape for ProcessAssetExtractor (project = .Detection, namespace =
  Mithril.MapCalibration.Internal) and the rename CaptureJsonContext ->
  DetectionJsonContext (absorbing 5 [JsonSerializable] entries).
- MapCalibrationLocateOptions namespace did change from .Capture to .Detection
  (Detection-tier concern, not Capture); all 5 test consumers updated.

Spec: docs/planning/map-calibration-detection-project-split/spec.md
