+++ src/Mithril.MapCalibration.Detection/BorderMask.cs
