+++ src/Mithril.MapCalibration.Detection/CalibrationConfidenceGate.cs
