+++ src/Mithril.MapCalibration.Detection/DeviationBlobCalibrationDetector.cs
