+++ src/Mithril.MapCalibration.Detection/DeviationBlobDetector.cs
