+++ src/Mithril.MapCalibration.Detection/DeviationFloodRimMask.cs
