+++ src/Mithril.MapCalibration.Detection/GrayImage.cs
