+++ src/Mithril.MapCalibration.Detection/IBaseTextureProvider.cs
