+++ src/Mithril.MapCalibration.Detection/ICalibrationDetector.cs
