+++ src/Mithril.MapCalibration.Detection/IIconTemplateProvider.cs
