+++ src/Mithril.MapCalibration.Detection/IMapRegionRefiner.cs
