+++ src/Mithril.MapCalibration.Detection/IconLikelihoodField.cs
