+++ src/Mithril.MapCalibration.Detection/IconRenderScaler.cs
