+++ src/Mithril.MapCalibration.Detection/IconTemplate.cs
