+++ src/Mithril.MapCalibration.Detection/IconTemplateCache.cs
