+++ src/Mithril.MapCalibration.Detection/IconTemplateSet.cs
