+++ src/Mithril.MapCalibration.Detection/ImageOps.cs
