+++ src/Mithril.MapCalibration.Detection/Internal/CachedIconTemplateProvider.cs
