+++ src/Mithril.MapCalibration.Detection/Internal/CanonicalAssetHashes.cs
