+++ src/Mithril.MapCalibration.Detection/Internal/IconTemplateManifest.cs
