+++ src/Mithril.MapCalibration.Detection/Internal/MapTextureManifest.cs
