+++ src/Mithril.MapCalibration.Detection/JEvaluator.cs
