+++ src/Mithril.MapCalibration.Detection/LocalNccDeviation.cs
