+++ src/Mithril.MapCalibration.Detection/MapCalibrationSolveEngine.cs
