+++ src/Mithril.MapCalibration.Detection/NccTemplateMatch.cs
