+++ src/Mithril.MapCalibration.Detection/RimMaskMode.cs
