+++ src/Mithril.MapCalibration.Detection/TypeAwareRansacSolver.cs
