+++ src/Mithril.MapCalibration.Detection/TypedDetection.cs
