+++ src/Mithril.MapCalibration.Detection/WholeImageTemplateDetector.cs
