+++ src/Mithril.MapCalibration/DependencyInjection/MapCalibrationServiceCollectionExtensions.cs
@@ -1,9 +1,6 @@
 using System.IO;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
-using Mithril.MapCalibration.Detection;
-using Mithril.MapCalibration.Detection.Internal;
 using Mithril.MapCalibration.Internal;
 
 namespace Mithril.MapCalibration.DependencyInjection;
@@ -61,62 +58,4 @@ public static class MapCalibrationServiceCollectionExtensions
         return new MapCalibrationService(baseline, store, goodResidualThresholdPx, serviceLogger);
     }
 
-    /// <summary>
-    /// Register the headless detect→solve engine (Phase 1): the deviation-blob
-    /// <see cref="ICalibrationDetector"/>, the
-    /// <see cref="ICalibrationConfidenceGate"/>, the
-    /// <see cref="MapCalibrationSolveEngine"/>, the
-    /// <see cref="IIconTemplateProvider"/> (a per-attempt
-    /// <see cref="Internal.CachedIconTemplateProvider"/> over the asset cache dir),
-    /// and an <see cref="IBaseTextureProvider"/> over the same cache. Independent of
-    /// <see cref="AddMithrilMapCalibration"/> (the persistence registration) —
-    /// register either or both.
-    ///
-    /// <para><b>#931:</b> the icon templates + base textures are no longer shipped
-    /// as embedded PG art; the out-of-process asset-extractor sidecar populates
-    /// <paramref name="assetCacheDir"/> at runtime and these loaders read it
-    /// BCL-only. When the dir is absent/empty the provider yields
-    /// <see cref="IconTemplateSet.Empty"/> / the base-texture provider returns
-    /// null — the intended fail-soft (no detections → gate rejects →
-    /// safe-degrade). The optional <paramref name="pgVersion"/> keys the
-    /// canonical-hash gate that guards base textures.</para>
-    ///
-    /// <para><b>#949:</b> the icon templates resolve via a <i>per-attempt</i>
-    /// provider (re-reads the cache each call), not an eager singleton, so a
-    /// same-session populate (the engine's <c>--icons</c> demand-trigger or the
-    /// startup warm-up) engages on the next attempt — first-session calibration no
-    /// longer needs a restart.</para>
-    /// </summary>
-    public static IServiceCollection AddMithrilMapCalibrationEngine(
-        this IServiceCollection services,
-        string assetCacheDir,
-        string? pgVersion = null)
-    {
-        if (string.IsNullOrWhiteSpace(assetCacheDir))
-            throw new ArgumentException("assetCacheDir required", nameof(assetCacheDir));
-
-        services.AddSingleton<IIconTemplateProvider>(sp =>
-            new CachedIconTemplateProvider(
-                assetCacheDir,
-                sp.GetService<ILoggerFactory>()?.CreateLogger("Mithril.MapCalibration.Templates")));
-        services.AddSingleton<IBaseTextureProvider>(sp =>
-        {
-            var loggerFactory = sp.GetService<ILoggerFactory>();
-            var gate = CanonicalAssetHashGate.Load(loggerFactory?.CreateLogger("Mithril.MapCalibration.HashGate"));
-            return new CachedBaseTextureProvider(
-                assetCacheDir,
-                gate,
-                pgVersion,
-                loggerFactory?.CreateLogger("Mithril.MapCalibration.BaseTexture"));
-        });
-        services.AddSingleton<ICalibrationDetector, DeviationBlobCalibrationDetector>();
-        services.AddSingleton<ICalibrationConfidenceGate, CalibrationConfidenceGate>();
-        services.TryAddSingleton<MapCalibrationSolverOptions>();
-        services.AddSingleton(sp => new MapCalibrationSolveEngine(
-            sp.GetRequiredService<ICalibrationDetector>(),
-            sp.GetRequiredService<ICalibrationConfidenceGate>(),
-            sp.GetService<ILoggerFactory>()?.CreateLogger("Mithril.MapCalibration.Engine"),
-            sp.GetRequiredService<MapCalibrationSolverOptions>()));
-        return services;
-    }
 }
