+++ src/Mithril.MapCalibration/Internal/MapCalibrationJsonContext.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
-using Mithril.MapCalibration.Detection.Internal;
 
 namespace Mithril.MapCalibration.Internal;
 
@@ -26,10 +25,6 @@ namespace Mithril.MapCalibration.Internal;
 [JsonSerializable(typeof(UserRefinementFile))]
 [JsonSerializable(typeof(Dictionary<string, AreaCalibration>))]
 [JsonSerializable(typeof(AreaCalibration))]
-[JsonSerializable(typeof(IconTemplateManifest))]
-[JsonSerializable(typeof(MapTextureManifest))]
-[JsonSerializable(typeof(CanonicalAssetHashes))]
-[JsonSerializable(typeof(SidecarResult))]
 internal sealed partial class MapCalibrationJsonContext : JsonSerializerContext;
 
 /// <summary>On-disk shape for the embedded baseline JSON resource.</summary>
