+++ tools/Mithril.MapCalibration.Tools.Common/Mithril.MapCalibration.Tools.Common.csproj
@@ -21,5 +21,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Mithril.MapCalibration\Mithril.MapCalibration.csproj" />
+    <ProjectReference Include="..\..\src\Mithril.MapCalibration.Detection\Mithril.MapCalibration.Detection.csproj" />
   </ItemGroup>
 </Project>
